fix(web): Rework the first time modal tutorial experience
This wasn't working in App.vue so we needed to move it to WorkspaceSinglePage - this is guaranteed to be an authenticated user and we are using localstorage to control that we don't serve the popup on a page refresh
