+++ app/web/src/pages/WorkspaceSinglePage.vue
@@ -72,17 +72,57 @@
           class="flex-none"
         />
       </template>
+      <Modal
+        ref="firstTimeModalRef"
+        noExit
+        size="2xl"
+        title="Welcome To System Initiative!"
+      >
+        <!-- TODO(Wendy) - PLACEHOLDER VIDEO, please replace with our video before pushing to prod! -->
+        <iframe
+          class="aspect-video"
+          src="https://www.youtube.com/embed/7vrIJmP49IE?si=Kknr-Qm5DDBDXjTu"
+          title="YouTube video player"
+          frameborder="0"
+          allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
+          referrerpolicy="strict-origin-when-cross-origin"
+          allowfullscreen
+        ></iframe>
+        <div class="flex flex-row gap-sm mt-xs">
+          <VormInput
+            v-model="firstTimeModalCheckbox"
+            class="flex flex-row-reverse gap-0 italic"
+            type="checkbox"
+            label="Don't show me this video again."
+            inlineLabel
+          />
+          <VButton
+            class="flex-grow"
+            label="Let's Get Started!"
+            @click="closeFirstTimeModal"
+          />
+        </div>
+      </Modal>
     </template>
   </AppLayout>
 </template>
 
 <script lang="ts" setup>
-import { computed, PropType, watch } from "vue";
+import { computed, onMounted, PropType, ref, watch } from "vue";
 import { useRoute, useRouter } from "vue-router";
 import * as _ from "lodash-es";
-import { ErrorMessage, Icon, LoadingMessage } from "@si/vue-lib/design-system";
+import storage from "local-storage-fallback";
+import {
+  ErrorMessage,
+  Icon,
+  LoadingMessage,
+  Modal,
+  VButton,
+  VormInput,
+} from "@si/vue-lib/design-system";
 import { useChangeSetsStore } from "@/store/change_sets.store";
 import { useWorkspacesStore } from "@/store/workspaces.store";
+import { useAuthStore } from "@/store/auth.store";
 import AppLayout from "@/components/layout/AppLayout.vue";
 import Navbar from "@/components/layout/navbar/Navbar.vue";
 import StatusBar from "@/components/StatusBar/StatusBar.vue";
@@ -96,6 +136,7 @@ const route = useRoute();
 
 const workspacesStore = useWorkspacesStore();
 const changeSetsStore = useChangeSetsStore();
+const authStore = useAuthStore();
 
 const workspacesReqStatus = workspacesStore.getRequestStatus(
   "FETCH_USER_WORKSPACES",
@@ -108,6 +149,29 @@ const changeSetsReqStatus =
 
 // this page is the parent of many child routes so we watch the route rather than use mounted hooks
 watch([route, changeSetsReqStatus], handleUrlChange, { immediate: true });
+
+const firstTimeModalFired = ref(false);
+const firstTimeModalRef = ref<InstanceType<typeof Modal>>();
+onMounted(async () => {
+  if (authStore.user) {
+    const showModal = await authStore.CHECK_FIRST_MODAL(authStore.user.pk);
+
+    const hasServedModal = storage.getItem("SI_FIRST_TIME_MODAL_SHOWN");
+    if (!firstTimeModalFired.value && showModal && !hasServedModal) {
+      firstTimeModalRef.value?.open();
+      firstTimeModalFired.value = true;
+      storage.setItem("SI_FIRST_TIME_MODAL_SHOWN", "1");
+    }
+  }
+});
+
+const firstTimeModalCheckbox = ref(false);
+const closeFirstTimeModal = () => {
+  if (authStore.user && firstTimeModalCheckbox.value) {
+    authStore.DISMISS_FIRST_TIME_MODAL(authStore.user.pk);
+  }
+  firstTimeModalRef.value?.close();
+};
 
 // TODO: this logic needs some work
 function handleUrlChange() {
